requirements.txt
Flask==3.0.2
Flask-Login==0.6.3
Flask-WTF==1.2.1
email-validator==2.1.1
PyMySQL==1.1.1
SQLAlchemy==2.0.29
python-dotenv==1.0.1
